Remove duplicates and move to a helper file
